// web/src/gen/model/filterField.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */
import type { FilterOperator } from "./filterOperator";

export interface FilterField {
  field: string;
  operator: FilterOperator;
  value: string;
}